Fix overrunning channels
The array `source_list_s::channels` holds the channels required by the
associated source.
In the prior implementation, the loops expect the array is terminated by
-1. However, it was actually not.

This is a bug but fortunately `channels` is allocated by `bzalloc` and
the number of channels is currently fixed 2 so that this won't cause
memory access violations but just accessing the channel-0.

+++ src/capdev-internal.h
@@ -7,7 +7,7 @@ struct source_list_s
 {
 	source_t *src;
 	uint64_t channel_mask;
-	size_t n_channels;
+	uint32_t n_channels;
 	int channels[N_CHANNELS];
 
 	struct source_list_s *next;
